+++ bs/sbox-forward.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-void sbox_forward(const uint8_t *inputs, uint8_t *outputs);
+void sbox_forward(const uint16_t *inputs, uint16_t *outputs);
